add fake datastore in datastore package
