+++ pkg/datastore/fakedatastore.go
@@ -1,0 +1,40 @@
+package datastore
+
+import (
+	"sync"
+
+	"github.com/llm-d/llm-d-inference-payload-processor/pkg/framework/datalayer"
+)
+
+// fakeDataStore is an in-memory DataStore for tests.
+type fakeDataStore struct {
+	mu     sync.Mutex
+	models map[string]datalayer.Model
+}
+
+func NewFakeDataStore() *fakeDataStore {
+	return &fakeDataStore{models: make(map[string]datalayer.Model)}
+}
+
+func (f *fakeDataStore) GetOrCreateModel(name string) datalayer.Model {
+	f.mu.Lock()
+	defer f.mu.Unlock()
+	if m, ok := f.models[name]; ok {
+		return m
+	}
+	m := datalayer.NewModel(name)
+	f.models[name] = m
+	return m
+}
+
+func (f *fakeDataStore) DeleteModel(name string) {}
+
+func (f *fakeDataStore) Models() []string {
+	f.mu.Lock()
+	defer f.mu.Unlock()
+	names := make([]string, 0, len(f.models))
+	for name := range f.models {
+		names = append(names, name)
+	}
+	return names
+}
